Update README.md
# Image-Algorithm-of-Paper
This repository is sustained by Jiayu Zhang and leaded by Kunjie Chen professor, both in Engineering Institution of NJAU University in China.
This repository has two parts, the one is pattern recognition and the other is artificial intelligence, which will implement in python or other OOP codes. We want to devote ourself to promote agriclulture devolopment which include plant and food phenotypic. 
We will make great efforts to ensure and improve the quality of this repository which maybe the initial goal of this project.